+++ offchain/crates/solana-sdk/src/shred_subscription/env.rs
@@ -7,3 +7,9 @@ pub mod development {
     pub const USDC_MINT_KEY: solana_sdk::pubkey::Pubkey =
         solana_sdk::pubkey!("uSDZq2RMuxrEf7gqgDjR8wJCtCyaDAQk2e5jLAaoeeM");
 }
+
+pub mod solana_devnet {
+    // Circle's USDC mint on Solana devnet, Crossmint uses this one.
+    pub const USDC_MINT_KEY: solana_sdk::pubkey::Pubkey =
+        solana_sdk::pubkey!("4zMMC9srt5Ri5X14GAgXhaHii3GnPAEERYPJgZJDncDU");
+}